+++ ChangeLog.md
@@ -18,3 +18,15 @@
 - Added "utils.py"
 - Reworked "Builder" Class
 - Added functionality to build second layer of statue
+
+# 12/19/24 Version 0.4
+
+- Added "ImageParser" class
+    - Responsible for processing an image file
+- Added "dict.py"
+- Updated "utils.py"
+    - Added "ProcessTextures" function
+    - Added "GetDifference" function
+    - Added "GetBlock" function
+    - Added "ConvertToBlocks" function
+- Added docstrings to all current functions and classes